backend/requirements.txt
# SmartEdge AI Inspector — Backend Dependencies
# Run: pip install -r requirements.txt

# ── Core web framework ──────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic

# ── Image / AI ──────────────────────────────────────────────────────
pillow
numpy
opencv-python-headless
torch
torchvision

# ── Gemini / Google AI ──────────────────────────────────────────────
google-generativeai
google-genai

# ── Google Auth (token verification) ────────────────────────────────
google-auth
requests

# ── HTTP client (used for ElevenLabs TTS proxy) ──────────────────────
httpx

# ── SSL cert bundle — fixes macOS dev SSL issues ─────────────────────
certifi

# ── Reporting ────────────────────────────────────────────────────────
reportlab
